+++ tests/open_alchemy/models_file/model/test_type.py
@@ -29,8 +29,8 @@ from open_alchemy import models_file
         ("string", "date", False, None, None, None, "datetime.date"),
         ("string", "date-time", False, None, None, None, "datetime.datetime"),
         ("boolean", None, False, None, None, None, "bool"),
-        ("object", None, False, None, None, "RefModel", '"RefModel"'),
-        ("array", None, None, None, None, "RefModel", 'typing.Sequence["RefModel"]'),
+        ("object", None, False, None, None, "RefModel", '"TRefModel"'),
+        ("array", None, None, None, None, "RefModel", 'typing.Sequence["TRefModel"]'),
         ("integer", None, None, None, None, None, "typing.Optional[int]"),
         ("integer", None, None, True, None, None, "int"),
         ("integer", None, None, None, True, None, "int"),
@@ -444,7 +444,7 @@ def test_model_database_type_many_to_one(engine, sessionmaker):
     assert queried_models[1].ref_table is None
 
     # Check that returned type is correct
-    assert calculated_type_str == 'typing.Optional["RefTable"]'
+    assert calculated_type_str == 'typing.Optional["TRefTable"]'
 
     # Creating instance of ref_model without models
     ref_model_instance3 = ref_model(id=31, name="ref table name 3")
@@ -469,7 +469,7 @@ def test_model_database_type_many_to_one(engine, sessionmaker):
     with pytest.raises(TypeError):
         ref_model(id=41, name="ref table name 4", tables=None)
 
-    assert calculated_backref_type_str == 'typing.Sequence["Table"]'
+    assert calculated_backref_type_str == 'typing.Sequence["TTable"]'
 
 
 @pytest.mark.models_file
@@ -529,7 +529,7 @@ def test_model_database_type_many_to_one_not_nullable(engine, sessionmaker):
         session.flush()
 
     # Check that returned type is correct
-    assert calculated_type_str == '"RefTable"'
+    assert calculated_type_str == '"TRefTable"'
 
 
 @pytest.mark.models_file
@@ -607,7 +607,7 @@ def test_model_database_type_one_to_one(engine, sessionmaker):
     assert queried_models[1].ref_table is None
 
     # Check that returned type is correct
-    assert calculated_type_str == 'typing.Optional["RefTable"]'
+    assert calculated_type_str == 'typing.Optional["TRefTable"]'
 
     # Creating instance of ref_model without model
     ref_model_instance3 = ref_model(id=31, name="ref table name 3")
@@ -624,7 +624,7 @@ def test_model_database_type_one_to_one(engine, sessionmaker):
     assert queried_ref_models[1].table is None
     assert queried_ref_models[2].table is not None
 
-    assert calculated_backref_type_str == 'typing.Optional["Table"]'
+    assert calculated_backref_type_str == 'typing.Optional["TTable"]'
 
 
 @pytest.mark.models_file
@@ -701,7 +701,7 @@ def test_model_database_type_one_to_one_not_nullable(engine, sessionmaker):
     assert queried_ref_models[0].table is None
     assert queried_ref_models[1].table is not None
 
-    assert calculated_backref_type_str == 'typing.Optional["Table"]'
+    assert calculated_backref_type_str == 'typing.Optional["TTable"]'
 
     # Creating models
     base.metadata.create_all(engine)
@@ -713,7 +713,7 @@ def test_model_database_type_one_to_one_not_nullable(engine, sessionmaker):
         session.flush()
 
     # Check that returned type is correct
-    assert calculated_type_str == '"RefTable"'
+    assert calculated_type_str == '"TRefTable"'
 
 
 @pytest.mark.models_file
@@ -801,7 +801,7 @@ def test_model_database_type_one_to_many(engine, sessionmaker):
     with pytest.raises(TypeError):
         model(id=41, name="ref table name 4", ref_tables=None)
 
-    assert calculated_type_str == 'typing.Sequence["RefTable"]'
+    assert calculated_type_str == 'typing.Sequence["TRefTable"]'
 
     # Creating instance of ref_model with model
     ref_model_instance5 = ref_model(
@@ -818,7 +818,7 @@ def test_model_database_type_one_to_many(engine, sessionmaker):
     assert queried_models[1].table is not None
     assert queried_models[2].table is None
 
-    assert calculated_backref_type_str == 'typing.Optional["Table"]'
+    assert calculated_backref_type_str == 'typing.Optional["TTable"]'
 
 
 @pytest.mark.models_file
@@ -906,7 +906,7 @@ def test_model_database_type_many_to_many(engine, sessionmaker):
     with pytest.raises(TypeError):
         model(id=41, name="ref table name 4", ref_tables=None)
 
-    assert calculated_type_str == 'typing.Sequence["RefTable"]'
+    assert calculated_type_str == 'typing.Sequence["TRefTable"]'
 
     # Creating instance of ref_model without models
     ref_model_instance5 = ref_model(id=51, name="ref table name 5")
@@ -931,4 +931,4 @@ def test_model_database_type_many_to_many(engine, sessionmaker):
     with pytest.raises(TypeError):
         ref_model(id=81, name="ref table name 8", tables=None)
 
-    assert calculated_backref_type_str == 'typing.Sequence["Table"]'
+    assert calculated_backref_type_str == 'typing.Sequence["TTable"]'